// src/app/share/models/activation-indicator.model.ts
export class ActivationIndicator {
  codePPD?: number;
  namePPD?: string;
  codeSupervisor?: number;
  nameSupervisor?: string;
  nameExpert?: string;
  nameIndicator: string;
  dateSeries: number[];
  series: number[];
  forecastSeries: number[];
  planValue: number;
  factValue: number;
  forecastValue: number;
}